+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.1.0
 *unreleased*
+* Add `dataToEsm` helper to create named exports from objects ([#29](https://github.com/rollup/rollup-pluginutils/issues/29))
 * Support literal keys in object patterns ([#27](https://github.com/rollup/rollup-pluginutils/issues/27))
 
 ## 2.0.1